refactor(mempool): state inferred DAG (#1234)
This refactors the mempool from semi-decoupled state, transaction, and batch DAGs, to a
single DAG. This simplifies the overall structure and is a more intuitive mempool model imo.

Nodes in the graph represent a transaction, batch or block and they are replaced as they are 
processed i.e. all transaction nodes from a batch will be replaced by that batch's node.

Edges in the graph are inferred via the state. Each node's produced state and consumed state
is tracked, and this let's us establish parent <-> child relations by connecting the producer and
consumer of a note, or account commitment. This acts as a level of indirection, such that we
can remove or replace a node without having to update all the parent/child pointers, since instead
these are computed on-demand.